# contrat-swagger.yml
openapi: 3.0.0
info:
  title: Students API
  description: Manage interaction between database and frontend application
  version: 1.0.0
paths:
  /students:
    get:
      summary: Get all students from database
      tags:
        - End points
      operationId: getStudents
      responses:
        "200":
          description: OK, returned all the students in database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getStudentsResponseBody200"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getStudentsResponseBody400"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getStudentsResponseBody404"
  /student:
    post:
      summary: Add a new student to database and autoincrement the id
      tags:
        - End points
      operationId: getStudents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postStudentRequestBody"
      responses:
        "201":
          description: Created, new student as been added to database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postStudentResponseBody201"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postStudentResponseBody400"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postStudentResponseBody500"
...
components:
   schemas:
...
      getStudentsResponseBody200:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      getStudentsResponseBody400:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      getStudentsResponseBody404:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      ##########################################################################
      # POST /student
      ##########################################################################
      postStudentRequestBody:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      postStudentResponseBody201:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      postStudentResponseBody400:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      postStudentResponseBody500:
        type: object
        properties:
          firstName:
            type: string
            minimum: 3
            maximum: 30
          lastName:
            type: string
            minimum: 3
            maximum: 30
          age:
            type: integer
            minimum: 1
            maximum: 3
          sex:
            type: string
            minimum: 1
            maximum: 10
      ##########################################################################
      # DELETE /studentId
      ##########################################################################
...